pleaseBack/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
frozenlist==1.7.0
git-filter-repo==2.47.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
stripe==12.4.0
twilio==9.7.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
